[PTT-331] window open
