+++ src/app/pages/avvisi-page/avvisi-step2-stato-caricamenti/avvisi-step2-stato-caricamenti.component.ts
@@ -6,6 +6,7 @@ import { UploadPaymentsService } from '../../../services/upload-payments.service
 import { TokenService } from '../../../services/token.service';
 import { PaymentJob } from '../../../models/payment-job';
 import { PaymentsService } from '../../../services/payments.service';
+import { DownloadService } from '../../../services/download.service';
 
 declare const $: any;
 
@@ -24,6 +25,7 @@ export class AvvisiStep2StatoCaricamentiComponent implements OnInit, OnDestroy {
   private interval = 0;
 
   constructor(
+    private downloadService: DownloadService,
     private router: Router,
     private uploadPaymentsService: UploadPaymentsService,
     private paymentsService: PaymentsService,
@@ -129,7 +131,9 @@ export class AvvisiStep2StatoCaricamentiComponent implements OnInit, OnDestroy {
 
   openDettagli(jobId?: number, fileName?: string): void {
     if (jobId && fileName) {
-      this.paymentsService.export(jobId, fileName);
+      this.paymentsService.export(jobId, fileName).subscribe(res => {
+        this.downloadService.downloadFile(res);
+      });
     }
   }
