+++ src/assets/csv/template.csv
@@ -1,3 +1,1 @@
 CodiceFiscale/P.IVA;Tipo Pagatore;Nome;Cognome;Indirizzo;Civico;Localita;Provincia;CAP;Nazione;Email;Telefono;ID Tenant;Importo;Causale
-PGPHZH83S05D089Z;F;4444444;t666666;Via Parigi;hhhhhh;jjjjjjj;PI;56123;IT;example@gmail.com;908098098;;234;/RFB/iuv/123/TXT/Descrizione Causale RPT
-15844561009;G;4444444;t666666;Via Parigi;hhhhhh;jjjjjjj;PI;56123;IT;example@gmail.com;908098098;;234;/RFB/iuv/89/TXT/Descrizione Causale RPT
